requirements.txt
# Core LlamaIndex - requires Pydantic V2
llama-index-core>=0.12.0

# LLM Integration - requires anthropic>=0.52.0
llama-index-llms-anthropic

# Embeddings
llama-index-embeddings-huggingface

# Vector Store
llama-index-vector-stores-chroma

# Pydantic - constrained to avoid schema generation errors
pydantic>=2.7.0,<2.10.0

# ChromaDB - latest version works with Pydantic 2
chromadb>=0.5.0

# Utilities
python-dotenv
sentence-transformers

# Unified llama-index (includes core + common integrations)
llama-index
